requirements.txt
Django==4.2
djangorestframework==3.16.0
requests==2.32.3 